Colorized output for -Q and -S
Complete colorized output for -Q and -S (hopefully).

A major refactoring was done with regards to how packages are printed.

enum dumplvl_t is introduced to indicate the appropriate level at which
to dump entire pacman databases to stdout.

pacman_db_info is renamed to pacman_db_dump to better reflect the purpose
of the function. pacman_db_info handles a NULL target list for:
-Q, -Qi, -Qs, -Si, -Ss
by dumping the relevant database(s).

which_db is a new helper function that returns a string indicating which
database a given package comes from. It can be either one of the sync
databases or "local".

print_pkg_pretty prints colorized output of -Q, -Qs and -Ss packages.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>