Multipurpose hash tables
This is the start of a series of patches to make hash tables effectively serve
multiple purposes, since we now have:
1. Normal hash tables
2. vindex (used in graph.c)

And a third type of hash table is imminent for use in dependency resolution.
Now, we have the normal hash table variant and vindex. To add a third type will
probably mean a lot of duplicate code when in fact, certain code can be shared
among the 3 hash tables. It's just that minor differences cause this integration
to become difficult.

So conventionally, we can:
1. Just create 3 different types of hash tables, with different files and funcs
2. Spray if statements all over the place (really ugly)

While analyzing this problem, I got this sudden revelation from Junio Hamano's
streaming API series on git.git:

1. an enum to classify the type of hash table
2. a union to store the different data types for the different hash tables
3. a function pointer table for the operations on the different tables
4. Expose a public hash table API which calls the specific functions previously
   setup in part 3

This is an extremely elegant solution. Currently, only the normal hash table
variant has been implemented.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>