Remove struct vindex and add VINDEX type for hash tables
Previously, our graph made use of struct vindex, which is a slightly modified
hash table.

struct vindex is removed and the graph makes use of a struct hash_table .
The hash_table is of VINDEX type and as such, uses VINDEX specific functions
which are newly introduced in hash.c . This set of functions are specially
tailored for hash tables to be used as an index in a graph.

The main purpose of this VINDEX type hash tables is to allow for fast location
queries (ie. determining where a piece of data lies in the adjacency list of
the graph).

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>