Data structure: hashmap
hashmap is a wrapper over the hash_table data type. It makes use of a HASH_BST
type hash_table internally and is an associative array which maps a single key
to a tree of values.

The main purpose of this data structure is to help in dependency resolution
involving provides.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>