Data structure: memlist
There are certain situations which require heavy use of pointers to a backing
data store, whose size can only be determined at run time. While linked lists
can do the job, there is heavy memory usage incurred due to the need to store
pointers for each node.

Arrays are resizable but we will need to manually keep track of the size of
the array and possibly realloc it, which is against our goal of having a data
store which will not have its memory address changed.

The memlist:
1. Allows us to have a backing data store which will not have its memory
   address changed (no reallocations)
2. Does all the tracking of size of memory pool and reallocation so we don't
   have to worry about that.

The memlist data structure is essentially a list of memory pools, one chained
after another. The memory pools in a memlist must store the same data type and
each pool is a fixed block of memory which is guaranteed to stay in the same
memory address throughout the duration of the program.

The memlist's behaviour is determined on its initialization and it behaves
differently when storing pointer data types.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>