Rename hashmap to hashbst
A hashmap is a key-value pair store. The current hashmap data structure maps a
single key to a tree of values and the name "hashmap" does not reflect the true
purpose of our data structure. While hashbst may not be a good name, it is at
least better than hashmap in reflecting the nature of the data structure.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>