Check for return value of normalize_package
Upon failure to download an AUR package, normalize_package returns NULL. In
crawl_resolve, the return value from the initial normalize_package is checked,
but not the ones when we normalize dependencies. They should also be checked
in the case of failure so we can abort the entire operation asap.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>