Preliminary -S functionality
The -S functionality which includes dependency resolution used by -Su and
--crawl is in the works. While this may seem like a mega patch, most of it
is actually removing legacy code from the former -S

Due to the change in -S code, many functions are removed. Most notably,
print_pkginfo is removed due to it not being used.

Previously, print_list took in a list of strings and a prefix as an argument
and proceeds to print the list of strings on the same line as the prefix, until
it ran out of lines, then continued the indenting print.

Now, we introduce a print function solely for printing lists until column space
runs out. It should be named print_list but print_list already exists. So
print_list is renamed to the more apt print_list_prefix and our new function
can use print_list as its name.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>