Add PW_LOG_VDEBUG log level
Debug messages are printed when PW_LOG_DEBUG is passed as the log level to the
pw_*printf functions, provided that the relevant bit is set in config->loglvl .
However, the debug messages are prefixed with "debug: ". This effectively
prevents the printing of debug messages broken into small pieces since there
will be unnecessary "debug: " prefixed before each piece of message.

The PW_LOG_VDEBUG log level is introduced to solve this problem. When the user
passes the --debug option on the command line, the bitmask in config->loglvl is
OR'ed with both PW_LOG_DEBUG and PW_LOG_VDEBUG. Use the PW_LOG_VDEBUG as the
log level when debug messages do not need / cannot be prefixed with "debug: ".

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>