Print immediate dependencies for -S
When installing new packages / checking outdated AUR packages, the entire
dependency graph is printed after dependency resolution. The graph can be huge
when syncing a lot of packages and as such, leads to unnecessary clutter.

Now, the dependency graph will only be printed when the --debug option is
supplied. Immediate dependencies will be printed regardless of the --debug
option. This gives the user a clearer view as to what will be sync'ed and
significantly reduces output.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>