Parse PKGBUILDs which use bash variable substitution
As highlighted in https://github.com/yanhan/powaur/issues/5

If a user tries to install a package with a PKGBUILD such as the following
(or if the package has a dependency whose PKGBUILD looks like the following):

	_pkg=liboil
	pkgname=lib32-$_pkg
	pkgver=0.3.17
	pkgrel=2
	pkgdesc="Library of simple functions that are optimized for various CPUs."
	arch=('x86_64')
	license=('custom')
	url="http://liboil.freedesktop.org/"
	depends=("$_pkg" 'lib32-glibc')

An erroneous sync will result. This is because powaur parses the depends array,
sees the tokens "$_pkg" and "lib32-glibc", then goes on to resolve dependencies
for packages "$_pkg" and "lib32-glibc". "$_pkg" is obviously an invalid package
and all hell breaks loose.

Teach parses_pkgbuild to perform a simple substitution of bash variables so
PKGBUILDs such as the one above which make use of bash substitution can be
parsed.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>